+++ Data Structures/sparsetable.cpp
@@ -8,7 +8,7 @@ private:
     vector<array<T, K + 1>> st;
     vector<int> log;
 public:
-    SparseTable(const vector<T>& a, auto op) : n(size(a)), op(op), st(n), log(n + 1) {
+    SparseTable(const vector<T>& a, auto op) : n((int)size(a)), op(op), st(n), log(n + 1) {
         for (int x = 1; x <= n; ++x) log[x] = __builtin_clz(1) - __builtin_clz(x);
         assert(log[n] <= K);
         for (int i = 0; i < n; ++i) st[i][0] = a[i];
