+++ String/kmp.cpp
@@ -1,5 +1,5 @@
 vector<int> prefix_function(const string& s) {
-    int n = size(s);
+    int n = (int)size(s);
     vector<int> p(n, 0);
     for (int i = 1; i < n; ++i) {
         int j = p[i - 1];
@@ -12,7 +12,7 @@ vector<int> prefix_function(const string& s) {
 
 // z-function from prefix function
 vector<int> z_function(const vector<int>& p) {
-    int n = size(p);
+    int n = (int)size(p);
     vector<int> z(n, 0);
     for (int i = 0; i < n; ++i) {
         int j = i - p[i] + 1;
@@ -30,7 +30,7 @@ vector<int> z_function(const vector<int>& p) {
 template<int K = 26, char offset = 'a'>
 vector<array<int, K>> compute_automaton(string s) {
     s += offset + K;
-    int n = size(s);
+    int n = (int)size(s);
     vector<int> p = prefix_function(s);
     vector<array<int, K>> aut(n);
     for (int i = 0; i < n; ++i)
